% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeoRun.R
\name{lap_table0}
\alias{lap_table0}
\title{Get Lap summary table from EpsonView object}
\usage{
lap_table0(.d)
}
\arguments{
\item{.d}{EpsonView_data object}
}
\description{
Get Lap summary table from EpsonView object
}
