% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeoRun.R
\name{NeoRun_data}
\alias{NeoRun_data}
\title{EpsonView_data() and NeoRun_data() returns dataframe of several data set as list,
including GPS data
Convert backup data to dataframe list}
\usage{
NeoRun_data(fname, runmemo = NULL)
}
\arguments{
\item{fname}{CSV file name}

\item{runmemo}{memo to add dataframe}
}
\description{
EpsonView_data() and NeoRun_data() returns dataframe of several data set as list,
including GPS data
Convert backup data to dataframe list
}
